requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
APScheduler==3.10.4
pandas
numpy
scikit-learn
joblib
scapy
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.1.8
SQLAlchemy==2.0.47
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.22
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
greenlet==3.1.1
